Move all modules to store\modules
