+++ src/app/store/modules/orders.js
@@ -1,6 +1,5 @@
 
 
-
 export default function (parent) {
   return {
     api: {url: '/orders',  parent: parent },
@@ -22,3 +21,26 @@ export default function (parent) {
   }
 
 };
+
+/**
+{
+   state() {
+       var properties = {};
+         for ( var i in APP_INSTANCE.design.AppFormOrder.children) {
+           properties[APP_INSTANCE.design.AppFormOrder.children[i]] = '';
+         }
+       return properties;
+   },
+
+   actions: {
+     makeOrder({dispatch, getters}, data) {
+       var config = {data:data};
+        config.method = 'POST';
+        config.url = getters.WEBSITE_API_URL + '/orders';
+        return  dispatch('callAPI', config, {root:true});
+
+     }
+   }
+
+}
+*/
