+++ src/app/store/modules/pages.js
