+++ src/core/store/appInstance.js
@@ -13,28 +13,7 @@ export default function ( APP_INSTANCE) {
        }
       return ;
     },
-   modules: {
-     order: {
-        state() {
-            var properties = {};
-              for ( var i in APP_INSTANCE.design.AppFormOrder.children) {
-                properties[APP_INSTANCE.design.AppFormOrder.children[i]] = '';
-              }
-            return properties;
-        },
-
-        actions: {
-          makeOrder({dispatch, getters}, data) {
-            var config = {data:data};
-             config.method = 'POST';
-             config.url = getters.WEBSITE_API_URL + '/orders';
-             return  dispatch('callAPI', config, {root:true});
-
-          }
-        }
-
-     }
-   },
+   
    mutations: {
 
       /**
